Add opt-in inflow gating, gate the staged CLI flows, widen the window accumulator to i128
counterparty_allowlist rules gain a gate_inflows flag (default false):
when set, KNOWN_ISSUER evaluates every leg of the value descriptor, so
receiving a token from a non-allowlisted issuer denies, including
asset-less inflow legs. Debit-leg behavior is unchanged in both modes.

The CLI pay and claim --sign-only and --submit-only flows now evaluate
operator policy under a v1 engine: the supplied envelope is decoded into
authoritative arguments with the same decoder the MCP commit phase uses
and sized through the canonical derivation, so policy evaluates what is
actually signed or broadcast. Undecodable envelopes follow the
opaque-signing posture with the allow_opaque_signing opt-in. The staged
flows match rules under the stellar_pay_commit and stellar_claim_commit
tool names; submit-only records the value_action_submitted audit row on
confirmed submit. Noop-engine behavior is unchanged.

The per-period rolling-window accumulator is i128-width end to end, so
cumulative recorded spend is exact across the full range; the fail-closed
refusal at the old i64 ceiling is replaced by exact accounting. The
store remains in-process state with no on-disk form.

Closes #39
Closes #40
Closes #42

+++ crates/stellar-agent-core/tests/policy_fixtures_adversarial.rs
@@ -372,7 +372,7 @@ fn seed_state_store(store: &PolicyStateStore, profile_name: &str, seeds: &[State
     for seed in seeds {
         let ts_ms = now_ms.saturating_sub(seed.offset_secs_ago.saturating_mul(1_000));
         store
-            .append(&key, ts_ms, seed.amount_stroops)
+            .append(&key, ts_ms, i128::from(seed.amount_stroops))
             .expect("state store append must succeed");
     }
 }
